New `replace:` feature for rendering
Sometimes, you need HTML entities that are a PITA to type out longhand.
For example, when I was writing my RCU posts[1], I often had to type
`&nbsp;&rarr;&nbsp;`, when it was easier to type ` -> `.

This commit introduces baked-in support for a `replace:` key in the
metadata for blog posts, pages and galleries.  It works like this:

    ---
    # normal header stuff...
    replace:
      " -> ": "&nbsp;&rarr;&nbsp;"

Then, any time the text ` -> ` appears in the body text, it will be
transformed into the appropriate sequence of escaped entities.

Fixes #2

[1]: https://jameshunt.us/writings/rcu.html