New `[% asset ... %]` directive, for sha1-renaming
This acts like a memoized content-sensitive file renamer.
The argument to `asset` is taken as relative to the `htdocs/` directory.

The first time we see a path to assetify, we read the file, perform a
deterministic one-way hash of it, and then rename the file to
incorporate that checksum into the filename.  This defeats caching CDNs
and the like.

Subsequent occurences of the same asset path will skip the file
operations.

All occurences are replaced with the `htdocs/`-relative path that
incorporates the content-aware checksum in the file name.